requirements.txt
# ── Core framework ────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9          # multipart/form-data file uploads

# ── LangChain + OpenAI ────────────────────────────────────────────────────────
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.6
openai>=1.30.0

# ── Vector store ──────────────────────────────────────────────────────────────
faiss-cpu>=1.8.0
numpy>=1.26.0

# ── Document parsing ─────────────────────────────────────────────────────────
pymupdf>=1.24.0                  # PDF parsing (fitz)
pypdf>=4.2.0                     # PDF fallback parser

# ── Database ─────────────────────────────────────────────────────────────────
sqlalchemy>=2.0.0

# ── Configuration ─────────────────────────────────────────────────────────────
pydantic-settings>=2.2.0
python-dotenv>=1.0.0

# ── Frontend ──────────────────────────────────────────────────────────────────
streamlit>=1.35.0
requests>=2.31.0
pandas>=2.2.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.2.0
pytest-asyncio>=0.23.0
httpx>=0.27.0                    # ASGI test client for FastAPI
